requirements.txt
# 费曼的猫 (Feynman's Cat) - 依赖列表

flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
openai>=1.0.0
